Rename apiError helper to api-error to match file naming.
