+++ frontend/src/stores/entryList.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
 import type { EntryListItemResponse } from '../../../types/api-responses'
-import { apiErrorMessage } from '../utils/apiError'
+import { apiErrorMessage } from '../utils/api-error'
 
 export const useEntryListStore = defineStore('entryList', () => {
   const entries = ref<EntryListItemResponse[]>([])
